requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
tensorflow-cpu==2.12.0
python-multipart
